# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "GO JWT-Auth",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
        "/auth/refresh": {
            "get": {
                "security": [
                    {
                        "Bearer-Access": []
                    },
                    {
                        "Bearer-Refresh": []
                    }
                ],
                "description": "Update access \u0026 refresh tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-in": {
            "get": {
                "description": "Get access \u0026 refresh tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign-in",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User-GUID",
                        "name": "guid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.TokenResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "string"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer-Access": {
            "description": "Type: Bearer YOUR_ACCESS_TOKEN",
            "type": "apiKey",
            "name": "Access",
            "in": "header"
        },
        "Bearer-Refresh": {
            "description": "Type: Bearer YOUR_REFRESH_TOKEN",
            "type": "apiKey",
            "name": "Refresh",
            "in": "header"
        }
    }
}